Fix/alur

+++ include/models/inventory/Inventory.hpp
@@ -45,7 +45,7 @@ class Inventory {
         int countAllBuildingsBasedOnPurchasePrice() const;
 
         bool isExistsTileBasedOnCode(std::string code) const;
-        std::map<std::string, std::vector<PropertyTile*>> getCompleteColourGroups(std::map<std::string, int> countTilesForEachColourBlock);
+        std::map<std::string, std::vector<PropertyTile*>> getCompleteColourGroups(std::map<std::string, size_t> countTilesForEachColourBlock);
         std::map<std::string, std::vector<PropertyTile*>> getOwnedPropertiesGroupByColourGroups();
 
         int countRailRoads() const;
@@ -54,7 +54,7 @@ class Inventory {
         int countProperty() const;
         int countCard() const;
 
-        std::map<std::string, std::vector<PropertyTile*>> groupPropertiesByColorGroup() const;  
+        std::map<std::string, std::vector<PropertyTile*>> groupPropertiesByColorGroup() const;
 };
